+++ conaconverter/converters/__init__.py
@@ -2,12 +2,14 @@ from .rekordbox import RekordboxReader, RekordboxWriter
 from .virtualdj import VirtualDjReader, VirtualDjWriter
 from .serato import SeratoReader, SeratoWriter
 from .engineos import EngineOsReader, EngineOsWriter
+from .traktor import TraktorReader, TraktorWriter
 
 READERS = {
     "rekordbox": RekordboxReader(),
     "serato":    SeratoReader(),
     "engineos":  EngineOsReader(),
     "virtualdj": VirtualDjReader(),
+    "traktor":   TraktorReader(),
 }
 
 WRITERS = {
@@ -15,6 +17,7 @@ WRITERS = {
     "serato":    SeratoWriter(),
     "engineos":  EngineOsWriter(),
     "virtualdj": VirtualDjWriter(),
+    "traktor":   TraktorWriter(),
 }
 
 FORMAT_LABELS = {
@@ -22,4 +25,5 @@ FORMAT_LABELS = {
     "serato":    "Serato",
     "engineos":  "Engine OS",
     "virtualdj": "VirtualDJ",
+    "traktor":   "Traktor NML",
 }